+++ src/model/bloodtyping/BloodTypingTest.java
@@ -16,7 +16,7 @@ public class BloodTypingTest {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="TINYINT")
+  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="SMALLINT")
   private Integer id;
 
   @Column(length=15)
